refactor: db.json, crew api 정리
