+++ db.json
@@ -1,205 +1,6 @@
 {
   "crew": [
     {
-      "title": "연습",
-      "content": "연습",
-      "crewname": "크루1",
-      "category": "러닝",
-      "image": "https://images.unsplash.com/photo-1679679008578-e3b6406bdb47?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=435&q=80",
-      "date": "2023-08-24",
-      "address": "서울 서초구 강남대로2길 3",
-      "detailAddress": "111호",
-      "id": 1
-    },
-    {
-      "title": "제목인 건 맞아요",
-      "content": "내용인 것도 맞아요",
-      "crewname": "크루명인 것도 맞구요",
-      "category": "러닝",
-      "image": "",
-      "date": "",
-      "address": "서울 서초구 강남대로2길 3",
-      "total": 50,
-      "detailAddress": "111",
-      "id": 2
-    },
-    {
-      "title": "연습",
-      "content": "연습",
-      "crewname": "크~루",
-      "category": "수영최고",
-      "image": "",
-      "date": "",
-      "address": "서울 서초구 강남대로2길 3",
-      "total": 50,
-      "detailAddress": "111",
-      "id": 3
-    },
-    {
-      "title": "제목이네요",
-      "content": "강아지는 귀엽다",
-      "crewname": "강아지산책모임",
-      "category": "강아지산책쓰",
-      "image": {},
-      "date": "",
-      "address": "서울 서초구 강남대로2길 3",
-      "total": 36,
-      "startDate": "2023-08-01T13:49:46.000Z",
-      "endDate": "2023-08-17T13:49:46.000Z",
-      "detailAddress": "111",
-      "id": 4
-    },
-    {
-      "title": "강아지는 귀엽군",
-      "content": "귀엽다니까",
-      "crewname": "강아지가 최고임",
-      "category": "강아지산책가자",
-      "image": {},
-      "date": "",
-      "address": "서울 서초구 강남대로2길 3",
-      "total": 8,
-      "startDate": "2023-08-01",
-      "endDate": "2023-08-25",
-      "detailAddress": "111",
-      "id": 5
-    },
-    {
-      "title": "머랭이",
-      "content": "머래니",
-      "crewname": "머래",
-      "category": "머래니데이",
-      "image": "",
-      "date": "",
-      "address": "서울 서초구 강남대로 27",
-      "total": 10,
-      "startDate": "2023-08-01",
-      "endDate": "2023-08-17",
-      "detailAddress": "111",
-      "id": 6
-    },
-    {
-      "title": "됐다! 대박",
-      "content": "ㄴㅇㄹ",
-      "crewname": "ㄴㅇㄹ",
-      "category": "대박이지 년월인이 된다니",
-      "image": {},
-      "date": "",
-      "address": "서울 서초구 강남대로2길 3",
-      "total": 8,
-      "startDate": "2023-02-01",
-      "endDate": "2023-08-17",
-      "detailAddress": "111",
-      "id": 7
-    },
-    {
-      "title": "오류가 사라졌나 볼까",
-      "content": "어디어디",
-      "crewname": "보자보자",
-      "category": "러닝",
-      "image": {},
-      "date": "",
-      "address": "서울 서초구 강남대로2길 3",
-      "total": 29,
-      "startDate": "2023-08-11",
-      "endDate": "2023-08-23",
-      "detailAddress": "111",
-      "id": 8
-    },
-    {
-      "title": "",
-      "content": "",
-      "crewname": "",
-      "category": "",
-      "image": "",
-      "address": "",
-      "total": "",
-      "startDate": "2023-08-03",
-      "endDate": "2023-08-17",
-      "id": 9
-    },
-    {
-      "title": "",
-      "content": "",
-      "crewname": "",
-      "category": "",
-      "image": "",
-      "address": "",
-      "total": "",
-      "startDate": "2023-08-03",
-      "id": 10
-    },
-    {
-      "title": "보여",
-      "content": "보이지?",
-      "crewname": "",
-      "category": "",
-      "image": "",
-      "address": "",
-      "total": "",
-      "startDate": "2023-08-02",
-      "endDate": "2023-08-03",
-      "id": 11
-    },
-    {
-      "title": "날짜최종연습",
-      "content": "최종이되면 좋겠다",
-      "crewname": "최종이되려나",
-      "category": "날짜최종보내기",
-      "image": "",
-      "address": "서울 서초구 강남대로2길 3",
-      "total": 8,
-      "dateRange": "2023-08-05 - 2023-08-11",
-      "detailAddress": "111",
-      "id": 12
-    },
-    {
-      "title": "오류가 뜰까 안 뜰까",
-      "content": "콘솔에서",
-      "crewname": "안뜨면 좋겠다",
-      "category": "날짜최종보내기",
-      "address": "서울 강남구 강남대로 238",
-      "total": 10,
-      "image": {},
-      "dateRange": "2023-08-05 - 2023-08-11",
-      "id": 13
-    },
-    {
-      "title": "다시",
-      "content": "다시",
-      "crewname": "다시",
-      "category": "분홍신",
-      "image": {},
-      "address": "서울 서초구 강남대로 27",
-      "total": 10,
-      "dateRange": "",
-      "detailAddress": "111",
-      "id": 14
-    },
-    {
-      "title": "",
-      "content": "",
-      "crewname": "",
-      "category": "",
-      "address": "서울 서초구 강남대로2길 6",
-      "total": "",
-      "image": {},
-      "id": 15
-    },
-    {
-      "title": "다시1",
-      "content": "다시1",
-      "crewname": "분홍신",
-      "category": "핑크슈즈",
-      "image": {},
-      "address": "서울 서초구 강남대로2길 6",
-      "total": 9,
-      "dateRange": "2023-08-10 - 2023-08-31",
-      "detailAddress": "111",
-      "id": 16
-    }
-  ],
-  "posts": [
-    {
       "id": 1,
       "image": "",
       "crewname": "youngburn",
