+++ src/api/TodoApi.jsx
@@ -52,9 +52,9 @@ export const useFetchTodos = () => {
   });
 };
 
-export const useFetchPosts = () => {
-  return useQuery(["posts"], async () => {
-    const { data } = await axios.get(`${BASE_URL}/posts`);
-    return data;
-  });
-};
+// export const useFetchPosts = () => {
+//   return useQuery(["posts"], async () => {
+//     const { data } = await axios.get(`${BASE_URL}/posts`);
+//     return data;
+//   });
+// };
